refactor: productsRepository was separated to productsRepo and productsQueryRepo
