+++ src/repositories/productsRepo.ts
@@ -1,26 +1,11 @@
 import { ProductViewModel } from "../models/products/ProductViewModel";
-import { getProductViewModel } from "../utils/getProductViewModel";
 import { ProductType, productsCollection } from "./db";
 
-export const productsRepository = {
-  async findProducts(
-    title: string | undefined | null
-  ): Promise<ProductViewModel[]> {
-    const filter: any = {};
-
-    if (title) {
-      filter.title = { $regex: title };
-    }
-
-    return productsCollection.find(filter).toArray();
-  },
-  async getProductById(id: number): Promise<ProductViewModel | null> {
-    let product: ProductType | null = await productsCollection.findOne({
-      id: id,
-    });
-
-    return product;
-  },
+/**
+ * This is the DAL (Data Access Layer).
+ * Which is responsible for CUD (CRUD without Read) operations.
+ */
+export const productsRepo = {
   async createProduct(newProduct: ProductType): Promise<ProductViewModel> {
     const result = await productsCollection.insertOne(newProduct);
     return newProduct;
